refactor: :recycle: refactor code
