+++ app.py
@@ -8,26 +8,8 @@ import subprocess
 from functools import wraps
 from chalice import Chalice
 
-# add src to the syspath
-from sys import path as syspath 
-from pathlib import Path
-from os.path import abspath
-
-def icludepath():
-    print('Cekcing configuration')
-    print(__file__)
-    cp = Path(__file__).parent
-    print(cp)
-    if abspath(cp) not in syspath:
-        print('CP added')
-        syspath.append(abspath(cp))
-    else:
-        print('Not added')
-    print('====')
-icludepath()
-# ---
-from config import Config # type:ignore
-from log import configlogging # type:ignore
+from src.config import Config # type:ignore
+from src.log import configlogging # type:ignore
 
 app = Chalice(app_name=Config.api_name)
