--- app/__init__.py
@@ -1,2 +1,0 @@
-__all__= ['app']
-from  .main import app
